+++ Application/EdFi.Ods.AdminApp.Web/ActionFilters/SetupRequiredFilter.cs
@@ -3,7 +3,6 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using System;
 #if NET48
 using System.Web.Mvc;
 #else
@@ -55,10 +54,5 @@ namespace EdFi.Ods.AdminApp.Web.ActionFilters
             var status = _getOdsStatusQuery.Execute(defaultInstanceName);
             return status != null && status == CloudOdsStatus.Ok;
         }
-    }
-
-    [AttributeUsage(AttributeTargets.Class)]
-    public class BypassSetupRequiredFilterAttribute : Attribute
-    {
     }
 }
